requirements.txt
nltk
torchtext==0.9